requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
requests==2.32.4
python-dotenv==1.1.1
apscheduler==3.11.0
SQLAlchemy==2.0.43
pydantic==2.11.7
akshare==1.16.71
